Add resizeable grid
The grid can be set to have any number of squares on each side and
it will size those squares accordingly so they all fit in the
container without changing the size of the container itself. Also
added mouseenter and mouseleave event listeners. These are empty
now but will eventually change the color of the square when the
cursor hovers over it.