Add new grid button, random colors, and darkening
The page will load with a 16x16 grid by default. If the user
clicks the "New Grid" button, the user will be prompted to
input how many squares on each side they want the new grid
to have. The user cannot enter a number greater than 100. The
current grid is deleted and a new one with the corresponding size
is created in its place.

As the user hovers the cursor over each square, the squares will
be colored in with a random color and will progressively get 10%
darker each time the user's cursor hovers over that square.