Add validation for new grid creation
The prompt that pops up when the user clicks the "New Grid" button
now ensures that a new grid is created ONLY if the input is a
number between 1 and 100. Anything else will alert the user that
they entered an invalid input. Clicking cancel or leaving the
input box blank will now cancel out of the prompt and leave the
old grid as is.